+++ app/src/main/java/com/sopt/yeogieottae/ui/compare/LikeListViewHolder.kt
@@ -1,4 +1,4 @@
-package com.sopt.yeogieottae.ui.compare.empty
+package com.sopt.yeogieottae.ui.compare
 
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
